Move game loop into class based approach with process method
